refs #1276   * [7523/old_svn] [7524/old_svn] [7525/old_svn] [7526/old_svn] [7527/old_svn] [7528/old_svn] [7529/old_svn] [7530/old_svn] [7531/old_svn] [7532/old_svn] [7533/old_svn] [7534/old_svn] [7535/old_svn] 를  1.8  트리에서 동작 가능하도록 수정하여 반영

+++ library/components/Needlworks.Cache.PageCache.php
@@ -142,8 +142,13 @@ class pageCache {
 		global $database;
 		if(doesHaveOwnership()) $this->_dbContents['owner'] = $this->dbContents;
 		else $this->_dbContents['user'] = $this->dbContents;
-		return POD::execute("REPLACE INTO {$database['prefix']}PageCacheLog 
-			VALUES(".getBlogId().", '".POD::escapeString($this->realName)."', '".POD::escapeString(serialize($this->_dbContents))."')");
+		$query = new TableQuery($database['prefix'].'PageCacheLog');
+		$query->setAttribute('blogid',getBlogId());
+		$query->setAttribute('name',POD::escapeString($this->realName),true);
+		$query->setAttribute('value',POD::escapeString(serialize($this->_dbContents)),true);
+		$query->setQualifier('blogid',getBlogId());
+		$query->setQualifier('name',POD::escapeString($this->realName),true);
+		return $query->replace();
 	}
 
 	private function removePageCacheLog() {
@@ -225,8 +230,15 @@ class queryCache {
 		if(!is_null($memcache)) {
 			return $memcache->set(getBlogId().'-'.$this->queryHash,serialize($this->contents));
 		} else {
-			return POD::execute("REPLACE INTO {$database['prefix']}PageCacheLog 
-				VALUES(".getBlogId().", '".POD::escapeString($this->queryHash)."', '".POD::escapeString(serialize($this->contents))."')");
+			$name = POD::escapeString($this->queryHash);
+			$value = POD::escapeString(serialize($this->contents));
+			$query = new TableQuery($database['prefix'].'PageCacheLog');
+			$query->setAttribute('blogid',getBlogId());
+			$query->setAttribute('name',$name,true);
+			$query->setAttribute('value',$value,true);
+			$query->setQualifier('blogid',getBlogId());
+			$query->setQualifier('name',$name,true);
+			return $query->replace();
 		}
 	}
 
@@ -261,14 +273,24 @@ class globalCacheStorage extends pageCache {
 	}
 	
 	function load() {
-		global $database;
+		global $database, $service;
+		if(isset($service['pagecache']) && $service['pagecache'] == false) return false;
 		$result = POD::queryCell("SELECT value FROM {$database['prefix']}PageCacheLog WHERE blogid = ".$this->_gBlogId." AND name = 'globalCacheStorage'");
 		if(isset($result)) $this->_gCacheStorage[$this->_gBlogId] = unserialize($result);
 	}
 
 	function save() {
-		global $database;
-		if($this->_isChanged) return POD::query("REPLACE INTO {$database['prefix']}PageCacheLog VALUES(".$this->_gBlogId.", 'globalCacheStorage', '".POD::escapeString(serialize($this->_gCacheStorage[$this->_gBlogId]))."')");
+		global $database, $service;
+		if(isset($service['pagecache']) && $service['pagecache'] == false) return false;
+		if($this->_isChanged) {	
+			$query = new TableQuery($database['prefix'].'PageCacheLog');
+			$query->setAttribute('blogid',$this->_gBlogId);
+			$query->setAttribute('name','globalCacheStorage',true);
+			$query->setAttribute('value',POD::escapeString(serialize($this->_gCacheStorage[$this->_gBlogId])),true);
+			$query->setQualifier('blogid',$this->_gBlogId);
+			$query->setQualifier('name','globalCacheStorage',true);
+			return $query->replace();
+		}
 	}
 	
 	function getContent($name) {
@@ -287,7 +309,8 @@ class globalCacheStorage extends pageCache {
 	}
 
 	function purge() {
-		global $database;
+		global $database, $service;
+		if(isset($service['pagecache']) && $service['pagecache'] == false) return false;
 		return POD::query("DELETE FROM {$database['prefix']}PageCacheLog WHERE blogid = ".$this->_gBlogId." AND name = 'globalCacheStorage'");
 	}
 }
